vmm_memory: Update the tool to work with latest Cloud Hypervisor
This patch improves the tool so that it can both work with latest
Cloud-Hypervisor and other VMM processes. It introduces two options
'--size=' and '--name=' that can't be used together. Depending on the
way the guest RAM region is exposed through smaps, one might want to
use one option or the other.

Signed-off-by: Sebastien Boeuf <sebastien.boeuf@intel.com>